[firebaseAuth] First implementation of authentication with Firebase
Remove deprecated methods and keep only registration and forgot password processes. The confirmation and password modification use Firebase screens with a specific UI. We will have to change that as we will no longer use our frontend to handle that.
Add `deleteAccount` feature. Need to check if it will successfully delete the potential referenced objects link to the account.